requirements.txt
playwright
python-dotenv
openpyxl
pandas
scipy
